Fix not printing |switch| as only element of loop body
A loop expected a jump(condition) at the start and assert if something
else popped up. A |switch| can be used to control the flow as well, so
something like wouldn't be printed at all, because the loop didn't know
what to do with the switch.

new a = 1;
while(true)
{
switch(a)
{
case 0:
{
// blah
}
default:
{
// blaah
return;
}
}
}