Fix private functions being listed twice in SM1.8
The compiler in SourceMod 1.8 introduced a way to use local functions as
callbacks without them being callable from outside the plugin itself.
This is done by adding all functions to the .publics table, but
prefixing the local ones with ".address.". This patch finds that pattern
and avoids having the function added twice, because the .dbg.symbols
section doesn't have that prefix in the name of the function.