Work around broken chained assignments in spcomp
There's been a bug in the SourcePawn compiler that let to garbage in `b` when doing `b = a = 5` instead of the intended value.
That has since been fixed in the compiler, but there are still binaries with that flaw floating around. Try to have it look right even though it won't work when executing.