Add experimental |goto| support
Print labels and |goto| statements.

`goto` was allowed pre SM 1.1 and still is in AMXX.

At the location of the label the compiler emits some instructions to reset the stack pointer in case of a goto from a different scope or skipping variable declarations. Detect that `lctrl + add.c + sctrl` pattern or later `stackkadjust` and mark that block as a target for a `goto`. Now check for any direct jumps to that block and replace it with a goto to that block instead of repeating it in place.