Try to print string literals in arguments when missing debug info
AMXX doesn't include the signature of natives, so try to guess if an argument might be a string and print it accordingly.

This can produce false positives especially when looking up global variables. In AMXX the addresses of global variables start at 0, so it'll replace all literal 0, 4, 8, 12 with the respective global variable. Haven't come up with some way to distinct them.