Fix printing dummy variables for by-ref parameters
When a function expects a parameter passed by reference, but the caller isn't interested in the value stored in that reference, the caller still has to provide a dummy variable for the call.

new value = 0;
if (GetTrieValue(hTrie, key, value)) ...

The |value| variable was falsely rewritten to be a the constant |0| like other stack temporaries, even though the constant 0 can't be loaded by a local reference expression.

Work around this problem by printing a placeholder name "_unused_temp"_" instead of trying to print 0. This needs more work to adjust for properly.

This flaw was hiding a backdoor in the plugin of the PainterHell challenge in JustCTF 2020 by cypis. Thank you for the challenge!