Try to avoid referencing from the middle of a string
Apply the same heuristic as for AMXX files where we require a null byte to be right before the referenced potential address. The idea is that it's the null terminator of the previous string.

If the string isn't preceded by another string this could fail, but it's better to detect less strings than a bunch of garbage.